@@ -1,37 +1,36 @@
 <?php
 
-namespace OpenApi\Examples\SwaggerSpec\Petstore;
-
-use OpenApi\Annotations as OA;
+namespace petstore;
 
 class PetsController
 {
+
     /**
-     * @OA\Get(
+     * @SWG\Get(
      *     path="/pets",
      *     summary="List all pets",
      *     operationId="listPets",
      *     tags={"pets"},
-     *     @OA\Parameter(
+     *     @SWG\Parameter(
      *         name="limit",
      *         in="query",
      *         description="How many items to return at one time (max 100)",
      *         required=false,
-     *         @OA\Schema(
-     *             type="integer",
-     *             format="int32"
-     *         )
+     *         type="integer",
+     *         format="int32"
      *     ),
-     *     @OA\Response(
+     *     @SWG\Response(
      *         response=200,
      *         description="An paged array of pets",
-     *         @OA\Schema(ref="#/components/schemas/Pets"),
-     *         @OA\Header(header="x-next", @OA\Schema(type="string"), description="A link to the next page of responses")
+     *         @SWG\Schema(ref="#/definitions/Pets"),
+     *         @SWG\Header(header="x-next", type="string", description="A link to the next page of responses")
      *     ),
-     *     @OA\Response(
+     *     @SWG\Response(
      *         response="default",
      *         description="unexpected error",
-     *         @OA\Schema(ref="#/components/schemas/Error")
+     *         @SWG\Schema(
+     *             ref="#/definitions/Error"
+     *         )
      *     )
      * )
      */
@@ -40,17 +39,17 @@ class PetsController
     }
 
     /**
-     * @OA\Post(
-     *     path="/pets",
-     *     summary="Create a pet",
-     *     operationId="createPets",
-     *     tags={"pets"},
-     *     @OA\Response(response=201, description="Null response"),
-     *     @OA\Response(
-     *         response="default",
-     *         description="unexpected error",
-     *         @OA\Schema(ref="#/components/schemas/Error")
-     *     )
+     * @SWG\Post(
+     *    path="/pets",
+     *    summary="Create a pet",
+     *    operationId="createPets",
+     *    tags={"pets"},
+     *    @SWG\Response(response=201, description="Null response"),
+     *    @SWG\Response(
+     *        response="default",
+     *        description="unexpected error",
+     *        @SWG\Schema(ref="#/definitions/Error")
+     *    )
      * )
      */
     public function createPets()
@@ -58,29 +57,27 @@ class PetsController
     }
 
     /**
-     * @OA\Get(
+     * @SWG\Get(
      *     path="/pets/{petId}",
      *     summary="Info for a specific pet",
      *     operationId="showPetById",
      *     tags={"pets"},
-     *     @OA\Parameter(
+     *     @SWG\Parameter(
      *         name="petId",
      *         in="path",
      *         required=true,
      *         description="The id of the pet to retrieve",
-     *         @OA\Schema(
-     *             type="string"
-     *         )
+     *         type="string"
      *     ),
-     *     @OA\Response(
+     *     @SWG\Response(
      *         response=200,
      *         description="Expected response to a valid request",
-     *         @OA\Schema(ref="#/components/schemas/Pets")
+     *         @SWG\Schema(ref="#/definitions/Pets")
      *     ),
-     *     @OA\Response(
+     *     @SWG\Response(
      *         response="default",
      *         description="unexpected error",
-     *         @OA\Schema(ref="#/components/schemas/Error")
+     *         @SWG\Schema(ref="#/definitions/Error")
      *     )
      * )
      */
