@@ -1,39 +1,24 @@
 <?php
-
-namespace OpenApi\Examples\UsingRefs;
-
-use OpenApi\Annotations as OA;
-
+namespace UsingRefs;
+?>
+A common scenario is to let swagger-php generate a definition based on your model class.
+These definitions can then be referenced with `ref="#/definitions/$classname"
+<?php
 /**
- * A common scenario is to let swagger-php generate a definition based on your model class.
- * These definitions can then be referenced with `ref="#/components/schemas/$classname".
+ * @SWG\Definition()
  */
+class Product {
 
-/**
- * @OA\Schema(
- *     description="Product model",
- *     type="object",
- *     title="Product model"
- * )
- */
-class Product extends Model
-{
     /**
      * The unique identifier of a product in our catalog.
      *
-     * @var int
-     *
-     * @OA\Property(format="int64", example=1)
+     * @var integer
+     * @SWG\Property(format="int64")
      */
     public $id;
 
     /**
-     * @OA\Property(ref="#/components/schemas/product_status")
+     * @SWG\Property(ref="#/definitions/product_status")
      */
     public $status;
-
-    /**
-     * @OA\Property
-     */
-    public StockLevel $stockLevel;
 }
