@@ -1,63 +1,46 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi\Annotations;
-
-use OpenApi\Generator;
+namespace Swagger\Annotations;
 
 /**
+ * @Annotation
  * Contact information for the exposed API.
  *
- * @see [OAI Contact Object](https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#contact-object)
- *
- * @Annotation
+ * A Swagger "Contact Object": https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#contactObject
  */
 class Contact extends AbstractAnnotation
 {
     /**
      * The identifying name of the contact person/organization.
-     *
      * @var string
      */
-    public $name = Generator::UNDEFINED;
+    public $name;
 
     /**
      * The URL pointing to the contact information.
-     *
      * @var string
      */
-    public $url = Generator::UNDEFINED;
+    public $url;
 
     /**
      * The email address of the contact person/organization.
-     *
      * @var string
      */
-    public $email = Generator::UNDEFINED;
+    public $email;
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_types = [
         'name' => 'string',
         'url' => 'string',
-        'email' => 'string',
+        'email' => 'string'
     ];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_parents = [
-        Info::class,
-    ];
-
-    /**
-     * @inheritdoc
-     */
-    public static $_nested = [
-        Attachable::class => ['attachables'],
+        'Swagger\Annotations\Info'
     ];
 }
