@@ -1,76 +1,55 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi\Annotations;
-
-use OpenApi\Generator;
+namespace Swagger\Annotations;
 
 /**
+ * @Annotation
  * Allows referencing an external resource for extended documentation.
  *
- * @see [OAI External Documentation Object](https://github.com/OAI/OpenAPI-Specification/blob/OpenAPI.next/versions/3.0.md#external-documentation-object)
- *
- * @Annotation
+ * A Swagger "External Documentation Object":  * https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#external-documentation-object
  */
 class ExternalDocumentation extends AbstractAnnotation
 {
     /**
      * A short description of the target documentation. GFM syntax can be used for rich text representation.
-     *
      * @var string
      */
-    public $description = Generator::UNDEFINED;
+    public $description;
 
     /**
      * The URL for the target documentation.
-     *
      * @var string
      */
-    public $url = Generator::UNDEFINED;
+    public $url;
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_types = [
         'description' => 'string',
         'url' => 'string',
     ];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_required = ['url'];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_parents = [
-        OpenApi::class,
-        Tag::class,
-        Schema::class,
-        AdditionalProperties::class,
-        Property::class,
-        Operation::class,
-        Get::class,
-        Post::class,
-        Put::class,
-        Delete::class,
-        Patch::class,
-        Head::class,
-        Options::class,
-        Trace::class,
-        Items::class,
-        JsonContent::class,
-        XmlContent::class,
-    ];
-
-    /**
-     * @inheritdoc
-     */
-    public static $_nested = [
-        Attachable::class => ['attachables'],
+        'Swagger\Annotations\Swagger',
+        'Swagger\Annotations\Tag',
+        'Swagger\Annotations\Schema',
+        'Swagger\Annotations\Definition',
+        'Swagger\Annotations\Property',
+        'Swagger\Annotations\Operation',
+        'Swagger\Annotations\Get',
+        'Swagger\Annotations\Post',
+        'Swagger\Annotations\Put',
+        'Swagger\Annotations\Delete',
+        'Swagger\Annotations\Patch',
+        'Swagger\Annotations\Head',
+        'Swagger\Annotations\Options',
+        'Swagger\Annotations\Items',
     ];
 }
