@@ -1,98 +1,72 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi\Annotations;
-
-use OpenApi\Generator;
+namespace Swagger\Annotations;
 
 /**
- * The object provides metadata about the API.
- *
- * The metadata may be used by the clients if needed and may be presented in editing or documentation generation tools for convenience.
- *
- * @see [OAI Info Object](https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#info-object)
- *
  * @Annotation
+ *
+ * A Swagger "Info Object":  https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#infoObject
  */
 class Info extends AbstractAnnotation
 {
     /**
      * The title of the application.
-     *
      * @var string
      */
-    public $title = Generator::UNDEFINED;
+    public $title;
 
     /**
-     * A short description of the application.
-     *
-     * CommonMark syntax may be used for rich text representation.
-     *
+     * A short description of the application. GFM syntax can be used for rich text representation.
      * @var string
      */
-    public $description = Generator::UNDEFINED;
+    public $description;
 
     /**
-     * An URL to the Terms of Service for the API.
-     *
-     * Must be in the format of an url.
-     *
+     * The Terms of Service for the API.
      * @var string
      */
-    public $termsOfService = Generator::UNDEFINED;
+    public $termsOfService;
 
     /**
      * The contact information for the exposed API.
-     *
      * @var Contact
      */
-    public $contact = Generator::UNDEFINED;
+    public $contact;
 
     /**
      * The license information for the exposed API.
-     *
      * @var License
      */
-    public $license = Generator::UNDEFINED;
+    public $license;
 
     /**
-     * The version of the OpenAPI document (which is distinct from the OpenAPI Specification version or the API implementation version).
-     *
+     * Provides the version of the application API (not to be confused by the specification version).
      * @var string
      */
-    public $version = Generator::UNDEFINED;
+    public $version;
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_required = ['title', 'version'];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_types = [
         'title' => 'string',
-        'version' => 'string',
         'description' => 'string',
-        'termsOfService' => 'string',
+        'termsOfService' => 'string'
     ];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_nested = [
-        Contact::class => 'contact',
-        License::class => 'license',
-        Attachable::class => ['attachables'],
+        'Swagger\Annotations\Contact' => 'contact',
+        'Swagger\Annotations\License' => 'license'
     ];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_parents = [
-        OpenApi::class,
+        'Swagger\Annotations\Swagger'
     ];
 }
