@@ -1,62 +1,57 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi\Annotations;
+namespace Swagger\Annotations;
+
+use Swagger\Logger;
 
 /**
- * The description of an item in a Schema with type `array`.
- *
  * @Annotation
+ * The description of an item in a Schema with type "array"
  */
 class Items extends Schema
 {
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_nested = [
-        Discriminator::class => 'discriminator',
-        Items::class => 'items',
-        Property::class => ['properties', 'property'],
-        ExternalDocumentation::class => 'externalDocs',
-        Xml::class => 'xml',
-        AdditionalProperties::class => 'additionalProperties',
-        Attachable::class => ['attachables'],
+        'Swagger\Annotations\Items' => 'items',
+        'Swagger\Annotations\Property' => ['properties', 'property'],
+        'Swagger\Annotations\ExternalDocumentation' => 'externalDocs',
+        'Swagger\Annotations\Xml' => 'xml'
     ];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_parents = [
-        Property::class,
-        AdditionalProperties::class,
-        Schema::class,
-        JsonContent::class,
-        XmlContent::class,
-        Items::class,
+        'Swagger\Annotations\Parameter',
+        'Swagger\Annotations\Property',
+        'Swagger\Annotations\Schema',
+        'Swagger\Annotations\Definition',
+        'Swagger\Annotations\Items',
+        'Swagger\Annotations\Header',
     ];
 
-    /**
-     * @inheritdoc
-     */
-    public function validate(array $stack = [], array $skip = [], string $ref = '', $context = null): bool
+    /** @inheritdoc */
+    public function validate($parents = [], $skip = [], $ref = '')
     {
         if (in_array($this, $skip, true)) {
             return true;
         }
-
-        $valid = parent::validate($stack, $skip, $ref, $context);
-
-        $parent = end($stack);
-        if ($parent instanceof Schema && $parent->type !== 'array') {
-            $this->_context->logger->warning('@OA\\Items() parent type must be "array" in ' . $this->_context);
-            $valid = false;
+        $valid = parent::validate($parents, $skip);
+        if (!$this->ref) {
+            $parent = end($parents);
+            if (is_object($parent) && ($parent instanceof Parameter && $parent->in !== 'body' || $parent instanceof Header)) {
+                // This is a "Items Object" https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#items-object
+                // A limited subset of JSON-Schema's items object.
+                $allowedTypes = ['string', 'number', 'integer', 'boolean', 'array'];
+                if (in_array($this->type, $allowedTypes) === false) {
+                    Logger::notice('@SWG\Items()->type="'.$this->type.'" not allowed inside a '.$parent->_identity([]).' must be "'.implode('", "', $allowedTypes).'" in ' . $this->_context);
+                    $valid = false;
+                }
+            }
         }
-
-        // @todo Additional validation when used inside a Header or Parameter context.
-
         return $valid;
+        // @todo Additional validation when used inside a Header or Parameter context.
     }
 }
