@@ -1,300 +1,250 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi\Annotations;
+namespace Swagger\Annotations;
 
-use OpenApi\Generator;
+use \Swagger\Logger;
 
 /**
+ * @Annotation
  * Describes a single operation parameter.
  *
- * A unique parameter is defined by a combination of a name and location.
- *
- * @see [OAA Parameter Object](https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#parameter-object)
- *
- * @Annotation
+ * A Swagger "Parameter Object": https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#parameterObject
  */
 class Parameter extends AbstractAnnotation
 {
     /**
-     * The relative or absolute path to the endpoint.
-     *
-     * @see [Using refs](https://swagger.io/docs/specification/using-ref/)
-     *
-     * @var string|class-string|object
-     */
-    public $ref = Generator::UNDEFINED;
-
-    /**
-     * The key into <code>Components::parameters</code> or <code>PathItem::parameters</code> array.
-     *
+     * $ref See http://json-schema.org/latest/json-schema-core.html#rfc.section.7
      * @var string
      */
-    public $parameter = Generator::UNDEFINED;
+    public $ref;
 
     /**
-     * The (case-sensitive) name of the parameter.
-     *
-     * If in is "path", the name field must correspond to the associated path segment from the path field in the Paths Object.
-     *
-     * If in is "header" and the name field is "Accept", "Content-Type" or "Authorization", the parameter definition shall be ignored.
-     * For all other cases, the name corresponds to the parameter name used by the in property.
-     *
+     * The key into Swagger->parameters or Path->parameters array.
      * @var string
      */
-    public $name = Generator::UNDEFINED;
+    public $parameter;
 
     /**
-     * The location of the parameter.
-     *
-     * Possible values are "query", "header", "path" or "cookie".
-     *
+     * The name of the parameter. Parameter names are case sensitive. If in is "path", the name field MUST correspond to the associated path segment from the path field in the Paths Object. See Path Templating for further information. For all other cases, the name corresponds to the parameter name used based on the in property.
      * @var string
      */
-    public $in = Generator::UNDEFINED;
+    public $name;
 
     /**
-     * A brief description of the parameter.
-     *
-     * This could contain examples of use.
-     *
-     * CommonMark syntax may be used for rich text representation.
-     *
+     * The location of the parameter. Possible values are "query", "header", "path", "formData" or "body".
      * @var string
      */
-    public $description = Generator::UNDEFINED;
+    public $in;
 
     /**
-     * Determines whether this parameter is mandatory.
-     *
-     * If the parameter location is "path", this property is required and its value must be true.
-     * Otherwise, the property may be included and its default value is false.
-     *
-     * @var bool
-     */
-    public $required = Generator::UNDEFINED;
-
-    /**
-     * Specifies that a parameter is deprecated and should be transitioned out of usage.
-     *
-     * @var bool
-     */
-    public $deprecated = Generator::UNDEFINED;
-
-    /**
-     * Sets the ability to pass empty-valued parameters.
-     *
-     * This is valid only for query parameters and allows sending a parameter with an empty value.
-     *
-     * Default value is false.
-     *
-     * If style is used, and if behavior is n/a (cannot be serialized), the value of allowEmptyValue shall be ignored.
-     *
-     * @var bool
-     */
-    public $allowEmptyValue = Generator::UNDEFINED;
-
-    /**
-     * Describes how the parameter value will be serialized depending on the type of the parameter value.
-     *
-     * Default values (based on value of in): for query - form; for path - simple; for header - simple; for cookie - form.
-     *
+     * A brief description of the parameter. This could contain examples of use. GFM syntax can be used for rich text representation.
      * @var string
      */
-    public $style = Generator::UNDEFINED;
+    public $description;
 
     /**
-     * When this is true, parameter values of type array or object generate separate parameters for each value of the array or key-value pair of the map.
-     *
-     * For other types of parameters this property has no effect.
-     *
-     * When style is form, the default value is true.
-     * For all other styles, the default value is false.
-     *
-     * @var bool
+     * Determines whether this parameter is mandatory. If the parameter is in "path", this property is required and its value MUST be true. Otherwise, the property MAY be included and its default value is false.
+     * @var boolean
      */
-    public $explode = Generator::UNDEFINED;
+    public $required;
 
     /**
-     * Determines whether the parameter value should allow reserved characters, as defined by RFC3986 :/?#[]@!$&'()*+,;= to be included without percent-encoding.
-     *
-     * This property only applies to parameters with an in value of query.
-     *
-     * The default value is false.
-     *
-     * @var bool
-     */
-    public $allowReserved = Generator::UNDEFINED;
-
-    /**
-     * The schema defining the type used for the parameter.
-     *
+     * The schema defining the type used for the body parameter.
      * @var Schema
      */
-    public $schema = Generator::UNDEFINED;
+    public $schema;
 
     /**
-     * Example of the media type.
-     *
-     * The example should match the specified schema and encoding properties if present.
-     * The example object is mutually exclusive of the examples object.
-     * Furthermore, if referencing a schema which contains an example, the example value shall override the example provided by the schema.
-     * To represent examples of media types that cannot naturally be represented in JSON or YAML, a string value can contain the example with escaping where necessary.
-     *
+     * The type of the parameter. Since the parameter is not located at the request body, it is limited to simple types (that is, not an object). The value MUST be one of "string", "number", "integer", "boolean", "array" or "file". If type is "file", the consumes MUST be either "multipart/form-data" or " application/x-www-form-urlencoded" and the parameter MUST be in "formData".
+     * @var string
+     */
+    public $type;
+
+    /**
+     * The extending format for the previously mentioned type. See Data Type Formats for further details.
+     * @var string
+     */
+    public $format;
+
+    /**
+     * Sets the ability to pass empty-valued parameters. This is valid only for either query or formData parameters and allows you to send a parameter with a name only or an empty value. Default value is false.
+     * @var boolean
+     */
+    public $allowEmptyValue;
+    
+    /**
+     * Required if type is "array". Describes the type of items in the array.
+     * @var \Swagger\Annotations\Items
+     */
+    public $items;
+
+    /**
+     * Determines the format of the array if type array is used. Possible values are: csv - comma separated values foo,bar. ssv - space separated values foo bar. tsv - tab separated values foo\tbar. pipes - pipe separated values foo|bar. multi - corresponds to multiple parameter instances instead of multiple values for a single instance foo=bar&foo=baz. This is valid only for parameters in "query" or "formData". Default value is csv.
+     * @var string
+     */
+    public $collectionFormat;
+
+    /**
+     * Sets a default value to the parameter. The type of the value depends on the defined type. See http://json-schema.org/latest/json-schema-validation.html#anchor101.
      * @var mixed
      */
-    public $example = Generator::UNDEFINED;
+    public $default = UNDEFINED;
 
     /**
-     * Examples of the media type.
-     *
-     * Each example should contain a value in the correct format as specified in the parameter encoding.
-     * The examples object is mutually exclusive of the example object.
-     * Furthermore, if referencing a schema which contains an example, the examples value shall override the example provided by the schema.
-     *
-     * @var array<string,Examples>
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor17.
+     * @var number
      */
-    public $examples = Generator::UNDEFINED;
+    public $maximum;
 
     /**
-     * A map containing the representations for the parameter.
-     *
-     * The key is the media type and the value describes it.
-     * The map must only contain one entry.
-     *
-     * @var array<MediaType>|JsonContent|XmlContent|Attachable
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor17.
+     * @var boolean
      */
-    public $content = Generator::UNDEFINED;
+    public $exclusiveMaximum;
 
     /**
-     * Path-style parameters defined by RFC6570.
-     *
-     * @see [RFC6570](https://tools.ietf.org/html/rfc6570#section-3.2.7)
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor21.
+     * @var number
      */
-    public $matrix = Generator::UNDEFINED;
+    public $minimum;
 
     /**
-     * Label style parameters defined by RFC6570.
-     *
-     * @see [RFC6570](https://tools.ietf.org/html/rfc6570#section-3.2.5)
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor21.
+     * @var boolean
      */
-    public $label = Generator::UNDEFINED;
+    public $exclusiveMinimum;
 
     /**
-     * Form style parameters defined by RFC6570.
-     *
-     * This option replaces collectionFormat with a csv (when explode is false) or multi (when explode is true) value from OpenAPI 2.0.
-     *
-     * @see [RFC6570](https://tools.ietf.org/html/rfc6570#section-3.2.8)
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor26.
+     * @var integer
      */
-    public $form = Generator::UNDEFINED;
+    public $maxLength;
 
     /**
-     * Simple style parameters defined by RFC6570.
-     *
-     * This option replaces collectionFormat with a csv value from OpenAPI 2.0.
-     *
-     * @see [RFC6570](https://tools.ietf.org/html/rfc6570#section-3.2.2)
-     *
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor29.
+     * @var integer
+     */
+    public $minLength;
+
+    /**
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor33.
+     * @var string
+     */
+    public $pattern;
+
+    /**
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor42.
+     * @var integer
+     */
+    public $maxItems;
+
+    /**
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor45.
+     * @var integer
+     */
+    public $minItems;
+
+    /**
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor49.
+     * @var boolean
+     */
+    public $uniqueItems;
+
+    /**
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor76.
      * @var array
      */
-    public $simple = Generator::UNDEFINED;
+    public $enum;
 
     /**
-     * Space separated array values.
-     *
-     * This option replaces collectionFormat equal to ssv from OpenAPI 2.0.
-     *
-     * @var array
+     * See http://json-schema.org/latest/json-schema-validation.html#anchor14.
+     * @var number
      */
-    public $spaceDelimited = Generator::UNDEFINED;
+    public $multipleOf;
 
-    /**
-     * Pipe separated array values.
-     *
-     * This option replaces collectionFormat equal to pipes from OpenAPI 2.0.
-     *
-     * @var array
-     */
-    public $pipeDelimited = Generator::UNDEFINED;
-
-    /**
-     * Provides a simple way of rendering nested objects using form parameters.
-     */
-    public $deepObject = Generator::UNDEFINED;
-
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_required = ['name', 'in'];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_types = [
         'name' => 'string',
-        'in' => ['query', 'header', 'path', 'cookie'],
+        'in' => ['query', 'header', 'path', 'formData', 'body'],
         'description' => 'string',
-        'style' => ['matrix', 'label', 'form', 'simple', 'spaceDelimited', 'pipeDelimited', 'deepObject'],
         'required' => 'boolean',
+        'format' => 'string',
+        'collectionFormat' => ['csv', 'ssv', 'tsv', 'pipes', 'multi'],
+        'maximum' => 'number',
+        'exclusiveMaximum' => 'boolean',
+        'minimum' => 'number',
+        'exclusiveMinimum' => 'boolean',
+        'maxLength' => 'integer',
+        'minLength' => 'integer',
+        'pattern' => 'string',
+        'maxItems' => 'integer',
+        'minItems' => 'integer',
+        'uniqueItems' => 'boolean',
+        'multipleOf' => 'integer',
     ];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_nested = [
-        Schema::class => 'schema',
-        Examples::class => ['examples', 'example'],
-        Attachable::class => ['attachables'],
+        'Swagger\Annotations\Items' => 'items',
+        'Swagger\Annotations\Schema' => 'schema'
     ];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_parents = [
-        Components::class,
-        PathItem::class,
-        Operation::class,
-        Get::class,
-        Post::class,
-        Put::class,
-        Delete::class,
-        Patch::class,
-        Head::class,
-        Options::class,
-        Trace::class,
+        'Swagger\Annotations\Operation',
+        'Swagger\Annotations\Get',
+        'Swagger\Annotations\Post',
+        'Swagger\Annotations\Put',
+        'Swagger\Annotations\Delete',
+        'Swagger\Annotations\Patch',
+        'Swagger\Annotations\Path',
+        'Swagger\Annotations\Head',
+        'Swagger\Annotations\Options',
+        'Swagger\Annotations\Swagger'
     ];
 
-    /**
-     * @inheritdoc
-     */
-    public function validate(array $stack = [], array $skip = [], string $ref = '', $context = null): bool
+    /** @inheritdoc */
+    public function validate($parents = [], $skip = [], $ref = '')
     {
         if (in_array($this, $skip, true)) {
             return true;
         }
-
-        $valid = parent::validate($stack, $skip, $ref, $context);
-
-        if (Generator::isDefault($this->ref)) {
+        $valid = parent::validate($parents, $skip, $ref);
+        if (empty($this->ref)) {
             if ($this->in === 'body') {
-                if (Generator::isDefault($this->schema)) {
-                    $this->_context->logger->warning('Field "schema" is required when ' . $this->identity() . ' is in "' . $this->in . '" in ' . $this->_context);
+                if ($this->schema === null) {
+                    Logger::notice('Field "schema" is required when ' . $this->identity() . ' is in "' . $this->in . '" in ' . $this->_context);
+                    $valid = false;
+                }
+            } else {
+                $validTypes = ['string', 'number', 'integer', 'boolean', 'array', 'file'];
+                if ($this->type === null) {
+                    Logger::notice($this->identity() . '->type is required when ' . $this->_identity([]) . '->in == "' . $this->in . '" in ' . $this->_context);
+                    $valid = false;
+                } elseif ($this->type === 'array' && $this->items === null) {
+                    Logger::notice($this->identity() . '->items required when ' . $this->_identity([]) . '->type == "array" in ' . $this->_context);
+                    $valid = false;
+                } elseif (in_array($this->type, $validTypes) === false) {
+                    $valid = false;
+                    Logger::notice($this->identity() . '->type must be "' . implode('", "', $validTypes) . '" when ' . $this->_identity([]) . '->in != "body" in ' . $this->_context);
+                } elseif ($this->type === 'file' && $this->in !== 'formData') {
+                    Logger::notice($this->identity() . '->in must be "formData" when ' . $this->_identity([]) . '->type == "file" in ' . $this->_context);
                     $valid = false;
                 }
             }
         }
-
         return $valid;
     }
 
-    /**
-     * @inheritdoc
-     */
-    public function identity(): string
+    /** @inheritdoc */
+    public function identity()
     {
         return parent::_identity(['name', 'in']);
     }
