@@ -1,12 +1,10 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi\Annotations;
-
-use OpenApi\Generator;
+namespace Swagger\Annotations;
 
 /**
  * @Annotation
@@ -15,33 +13,15 @@ class Property extends Schema
 {
     /**
      * The key into Schema->properties array.
-     *
      * @var string
      */
-    public $property = Generator::UNDEFINED;
+    public $property;
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_parents = [
-        AdditionalProperties::class,
-        Schema::class,
-        JsonContent::class,
-        XmlContent::class,
-        Property::class,
-        Items::class,
-    ];
-
-    /**
-     * @inheritdoc
-     */
-    public static $_nested = [
-        Discriminator::class => 'discriminator',
-        Items::class => 'items',
-        Property::class => ['properties', 'property'],
-        ExternalDocumentation::class => 'externalDocs',
-        Xml::class => 'xml',
-        AdditionalProperties::class => 'additionalProperties',
-        Attachable::class => ['attachables'],
+        'Swagger\Annotations\Definition',
+        'Swagger\Annotations\Schema',
+        'Swagger\Annotations\Property',
+        'Swagger\Annotations\Items',
     ];
 }
