@@ -1,138 +1,95 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi\Annotations;
-
-use OpenApi\Generator;
+namespace Swagger\Annotations;
 
 /**
- * @see [OAI Security Scheme Object](https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#securitySchemeObject).
- *
  * @Annotation
+ * A Swagger "Security Scheme Object": https://github.com/swagger-api/swagger-spec/blob/master/versions/2.0.md#securitySchemeObject
  */
 class SecurityScheme extends AbstractAnnotation
 {
     /**
-     * The relative or absolute path to a security scheme.
-     *
-     * @see [Using refs](https://swagger.io/docs/specification/using-ref/)
-     *
-     * @var string|class-string|object
-     */
-    public $ref = Generator::UNDEFINED;
-
-    /**
-     * The key into OpenApi->security array.
-     *
+     * The key into Swagger->securityDefinitions array.
      * @var string
      */
-    public $securityScheme = Generator::UNDEFINED;
+    public $securityDefinition;
 
     /**
      * The type of the security scheme.
-     *
      * @var string
      */
-    public $type = Generator::UNDEFINED;
+    public $type;
 
     /**
      * A short description for security scheme.
-     *
      * @var string
      */
-    public $description = Generator::UNDEFINED;
+    public $description;
 
     /**
      * The name of the header or query parameter to be used.
-     *
      * @var string
      */
-    public $name = Generator::UNDEFINED;
+    public $name;
 
     /**
      * Required The location of the API key.
-     *
      * @var string
      */
-    public $in = Generator::UNDEFINED;
+    public $in;
 
     /**
      * The flow used by the OAuth2 security scheme.
-     *
-     * @var Flow[]
+     * @var
      */
-    public $flows = Generator::UNDEFINED;
+    public $flow;
 
     /**
-     * A hint to the client to identify how the bearer token is formatted.
-     *
-     * Bearer tokens are usually generated by an authorization server, so this information is primarily for documentation purposes.
-     *
+     * The authorization URL to be used for this flow. This SHOULD be in the form of a URL.
      * @var string
      */
-    public $bearerFormat = Generator::UNDEFINED;
+    public $authorizationUrl;
 
     /**
-     * The name of the HTTP Authorization scheme.
-     *
-     * @see [RFC7235](https://tools.ietf.org/html/rfc7235#section-5.1)
-     *
+     * The token URL to be used for this flow. This SHOULD be in the form of a URL.
      * @var string
      */
-    public $scheme = Generator::UNDEFINED;
+    public $tokenUrl;
 
     /**
-     * OpenId Connect URL to discover OAuth2 configuration values. This MUST be in the form of a URL.
-     *
-     * @var string
+     * The available scopes for the OAuth2 security scheme.
+     * @var Scope[]
      */
-    public $openIdConnectUrl = Generator::UNDEFINED;
+    public $scopes;
 
-    /**
-     * @inheritdoc
-     */
-    public static $_required = ['securityScheme', 'type'];
+    /** @inheritdoc */
+    public static $_required = ['type'];
 
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_types = [
-        'type' => ['http', 'apiKey', 'oauth2', 'openIdConnect'],
+        'type' => ['basic', 'apiKey', 'oauth2'],
         'description' => 'string',
         'name' => 'string',
-        'bearerFormat' => 'string',
-        'in' => ['query', 'header', 'cookie'],
+        'in' => ['query', 'header'],
+        'flow' => ['implicit', 'password', 'application', 'accessCode']
     ];
 
-    /**
-     * @inheritdoc
-     */
-    public static $_nested = [
-        Flow::class => ['flows', 'flow'],
-        Attachable::class => ['attachables'],
-    ];
-
-    /**
-     * @inheritdoc
-     */
+    /** @inheritdoc */
     public static $_parents = [
-        Components::class,
+        'Swagger\Annotations\Swagger',
     ];
 
-    /**
-     * @inheritdoc
-     */
-    public function merge(array $annotations, bool $ignore = false): array
+    /** {@inheritdoc} */
+    #[\ReturnTypeWillChange]
+    public function jsonSerialize()
     {
-        $unmerged = parent::merge($annotations, $ignore);
-
-        if ($this->type === 'oauth2') {
-            $this->name = Generator::UNDEFINED;
+        if (is_array($this->scopes) && empty($this->scopes)) {
+            $this->scopes = new \StdClass();
         }
-
-        return $unmerged;
+        return parent::jsonSerialize();
     }
 }
