@@ -1,133 +1,106 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi;
-
-use OpenApi\Annotations as OA;
-use OpenApi\Loggers\DefaultLogger;
-use Psr\Log\LoggerInterface;
+namespace Swagger;
 
 /**
+ * Context
+ *
  * The context in which the annotation is parsed.
+ * It includes useful metadata which the Processors can use to augment the annotations.
  *
- * Contexts are nested to reflect the code/parsing hierarchy. They include useful metadata
- * which the processors can use to augment the annotations.
+ * Context hierarchy:
+ * - parseContext
+ *   |- docBlockContext
+ *   |- classContext
+ *      |- docBlockContext
+ *      |- propertyContext
+ *      |- methodContext
  *
- * @property string|null                           $comment     The PHP DocComment
- * @property string|null                           $filename
- * @property int|null                              $line
- * @property int|null                              $character
- * @property string|null                           $namespace
- * @property array|null                            $uses
- * @property string|null                           $class
- * @property string|null                           $interface
- * @property string|null                           $trait
- * @property string|null                           $enum
- * @property array|string|null                     $extends     Interfaces may extend a list of interfaces
- * @property array|null                            $implements
- * @property string|null                           $method
- * @property string|null                           $property
- * @property string|\ReflectionType|null           $type
- * @property bool|null                             $static      Indicate a static method
- * @property bool|null                             $nullable    Indicate a nullable value
- * @property bool|null                             $generated   Indicate the context was generated by a processor or
- *                                                              the serializer
- * @property Annotations\AbstractAnnotation|null   $nested
- * @property Annotations\AbstractAnnotation[]|null $annotations
- * @property LoggerInterface|null                  $logger      Guaranteed to be set when using the `Generator`
- * @property array|null                            $scanned     Details of file scanner when using ReflectionAnalyser
- * @property string|null                           $version     The OpenAPI version in use
+ * @property string $comment  The PHP DocComment
+ * @property string $filename
+ * @property int $line
+ * @property int $character
+ *
+ * @property string $namespace
+ * @property array $uses
+ * @property string $class
+ * @property string $extends
+ * @property string $method
+ * @property string $property
+ * @property Annotations\AbstractAnnotation[] $annotations
  */
-#[\AllowDynamicProperties]
 class Context
 {
     /**
      * Prototypical inheritance for properties.
-     *
-     * @var Context|null
+     * @var Context
      */
-    private $parent;
+    private $_parent;
 
-    public function clone()
-    {
-        return new Context(get_object_vars($this), $this->parent);
-    }
-
-    public function __construct(array $properties = [], ?Context $parent = null)
+    /**
+     * @param array $properties new properties for this context.
+     * @param Context $parent The parent context
+     */
+    public function __construct($properties = [], $parent = null)
     {
         foreach ($properties as $property => $value) {
-            $this->{$property} = $value;
+            $this->$property = $value;
         }
-        $this->parent = $parent;
-
-        $this->logger = $this->logger ?: new DefaultLogger();
+        $this->_parent = $parent;
     }
 
     /**
      * Check if a property is set directly on this context and not its parent context.
      *
-     * Example: $c->is('method') or $c->is('class')
+     * @param string $type Example: $c->is('method') or $c->is('class')
+     * @return bool
      */
-    public function is(string $property): bool
+    public function is($type)
     {
-        return property_exists($this, $property);
+        return property_exists($this, $type);
     }
 
     /**
-     * Check if a property is NOT set directly on this context and its parent context.
+     * Check if a property is NOT set directly on this context and but its parent context.
      *
-     * Example: $c->not('method') or $c->not('class')
+     * @param string $type Example: $c->not('method') or $c->not('class')
+     * @return bool
      */
-    public function not(string $property): bool
+    public function not($type)
     {
-        return $this->is($property) === false;
+        return property_exists($this, $type) === false;
     }
 
     /**
      * Return the context containing the specified property.
+     *
+     * @param string $property
+     * @return boolean|\Swagger\Context
      */
-    public function with(string $property): ?Context
+    public function with($property)
     {
-        if ($this->is($property)) {
+        if (property_exists($this, $property)) {
             return $this;
         }
-        if ($this->parent !== null) {
-            return $this->parent->with($property);
+        if ($this->_parent) {
+            return $this->_parent->with($property);
         }
-
-        return null;
+        return false;
     }
 
     /**
-     * Get the root context.
+     * @return \Swagger\Context
      */
-    public function root(): Context
+    public function getRootContext()
     {
-        if ($this->parent !== null) {
-            return $this->parent->root();
+        if ($this->_parent) {
+            return $this->_parent->getRootContext();
         }
-
         return $this;
-    }
-
-    /**
-     * Check if one of the given version numbers matches the current OpenAPI version.
-     *
-     * @param string|array $versions One or more version numbers
-     */
-    public function isVersion($versions): bool
-    {
-        if (!$this->version) {
-            throw new \RuntimeException('Version is only available reliably for validation and serialization');
-        }
-
-        $versions = (array) $versions;
-        $currentVersion = $this->version ?: OA\OpenApi::DEFAULT_VERSION;
-
-        return in_array($currentVersion, $versions);
     }
 
     /**
@@ -135,7 +108,7 @@ class Context
      *
      * @return string Example: "file1.php on line 12"
      */
-    public function getDebugLocation(): string
+    public function getDebugLocation()
     {
         $location = '';
         if ($this->class && ($this->method || $this->property)) {
@@ -161,19 +134,20 @@ class Context
                 $location .= ':' . $this->character;
             }
         }
-
         return $location;
     }
 
     /**
      * Traverse the context tree to get the property value.
+     *
+     * @param string $property
+     * @return mixed
      */
-    public function __get(string $property)
+    public function __get($property)
     {
-        if ($this->parent !== null) {
-            return $this->parent->{$property};
+        if ($this->_parent) {
+            return $this->_parent->$property;
         }
-
         return null;
     }
 
@@ -188,14 +162,86 @@ class Context
     }
 
     /**
-     * Create a Context based on the debug_backtrace.
-     *
-     * @deprecated
+     * A short piece of text, usually one line, providing the basic function of the associated element.
+     * @return string|null
      */
-    public static function detect(int $index = 0): Context
+    public function phpdocSummary()
     {
-        // trigger_deprecation('zircote/swagger-php', '4.0', 'Context detecting is deprecated');
+        $content = $this->phpdocContent();
+        if (!$content) {
+            return null;
+        }
+        $lines = preg_split('/(\n|\r\n)/', $content);
+        $summary = '';
+        foreach ($lines as $line) {
+            $summary .= $line."\n";
+            if ($line === '' || substr($line, -1) === '.') {
+                return trim($summary);
+            }
+        }
+        $summary = trim($summary);
+        if ($summary === '') {
+            return null;
+        }
+        return $summary;
+    }
+    
+    /**
+     * An optional longer piece of text providing more details on the associated element’s function. This is very useful when working with a complex element.
+     * @return string|null
+     */
+    public function phpdocDescription()
+    {
+        $summary = $this->phpdocSummary();
+        if (!$summary) {
+            return null;
+        }
+        $description = trim(substr($this->phpdocContent(), strlen($summary)));
+        if ($description === '') {
+            return null;
+        }
+        return $description;
+    }
 
+    /**
+     * The text contents of the phpdoc comment (excl. tags)
+     * @return string|null
+     */
+    public function phpdocContent()
+    {
+        $comment = preg_split('/(\n|\r\n)/', (string) $this->comment);
+        $comment[0] = preg_replace('/[ \t]*\\/\*\*/', '', $comment[0]); // strip '/**'
+        $i = count($comment) -1;
+        $comment[$i] = preg_replace('/\*\/[ \t]*$/', '', $comment[$i]); // strip '*/'
+        $lines = [];
+        $append = false;
+        foreach ($comment as $line) {
+            $line = ltrim($line, "\t *");
+            if (substr($line, 0, 1) === '@') {
+                break;
+            }
+            if ($append) {
+                $i = count($lines) - 1;
+                $lines[$i] = substr($lines[$i], 0, -1).$line;
+            } else {
+                $lines[] = $line;
+            }
+            $append = (substr($line, -1) === '\\');
+        }
+        $description = trim(implode("\n", $lines));
+        if ($description === '') {
+            return null;
+        }
+        return $description;
+    }
+
+    /**
+     * Create a Context based on the debug_backtrace
+     * @param int $index
+     * @return \Swagger\Context
+     */
+    public static function detect($index = 0)
+    {
         $context = new Context();
         $backtrace = debug_backtrace();
         $position = $backtrace[$index];
@@ -205,7 +251,7 @@ class Context
         if (isset($position['line'])) {
             $context->line = $position['line'];
         }
-        $caller = $backtrace[$index + 1] ?? null;
+        $caller = isset($backtrace[$index + 1]) ? $backtrace[$index + 1] : null;
         if (isset($caller['function'])) {
             $context->method = $caller['function'];
             if (isset($caller['type']) && $caller['type'] === '::') {
@@ -219,56 +265,50 @@ class Context
                 $context->namespace = implode('\\', $fqn);
             }
         }
-
         // @todo extract namespaces and use statements
         return $context;
     }
 
     /**
      * Resolve the fully qualified name.
+     *
+     * @param string $class  The class name
+     * @return string
      */
-    public function fullyQualifiedName(?string $source): string
+    public function fullyQualifiedName($class)
     {
-        if ($source === null) {
-            return '';
-        }
-
         if ($this->namespace) {
             $namespace = str_replace('\\\\', '\\', '\\' . $this->namespace . '\\');
         } else {
-            // global namespace
-            $namespace = '\\';
+            $namespace = '\\'; // global namespace
         }
-
-        $thisSource = $this->class ?? $this->interface ?? $this->trait;
-        if ($thisSource && strcasecmp($source, $thisSource) === 0) {
-            return $namespace . $thisSource;
+        if (strcasecmp($class, $this->class) === 0) {
+            return $namespace . $this->class;
         }
-        $pos = strpos($source, '\\');
+        $pos = strpos($class, '\\');
         if ($pos !== false) {
             if ($pos === 0) {
                 // Fully qualified name (\Foo\Bar)
-                return $source;
+                return $class;
             }
             // Qualified name (Foo\Bar)
             if ($this->uses) {
                 foreach ($this->uses as $alias => $aliasedNamespace) {
                     $alias .= '\\';
-                    if (strcasecmp(substr($source, 0, strlen($alias)), $alias) === 0) {
+                    if (strcasecmp(substr($class, 0, strlen($alias)), $alias) === 0) {
                         // Aliased namespace (use \Long\Namespace as Foo)
-                        return '\\' . $aliasedNamespace . substr($source, strlen($alias) - 1);
+                        return '\\' . $aliasedNamespace . substr($class, strlen($alias) - 1);
                     }
                 }
             }
         } elseif ($this->uses) {
             // Unqualified name (Foo)
             foreach ($this->uses as $alias => $aliasedNamespace) {
-                if (strcasecmp($alias, $source) === 0) {
+                if (strcasecmp($alias, $class) === 0) {
                     return '\\' . $aliasedNamespace;
                 }
             }
         }
-
-        return $namespace . $source;
+        return $namespace . $class;
     }
 }
