@@ -1,157 +1,111 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi\Processors;
+namespace Swagger\Processors;
 
-use OpenApi\Analysis;
-use OpenApi\Annotations as OA;
-use OpenApi\Context;
-use OpenApi\Generator;
+use Swagger\Annotations\Definition;
+use Swagger\Annotations\Items;
+use Swagger\Annotations\Property;
+use Swagger\Context;
+use Swagger\Analysis;
 
 /**
  * Use the property context to extract useful information and inject that into the annotation.
  */
-class AugmentProperties implements ProcessorInterface
+class AugmentProperties
 {
-    use Concerns\DocblockTrait;
-    use Concerns\RefTrait;
-    use Concerns\TypesTrait;
+    public static $types = [
+        'array' => 'array',
+        'byte' => ['string', 'byte'],
+        'boolean' => 'boolean',
+        'bool' => 'boolean',
+        'int' => 'integer',
+        'integer' => 'integer',
+        'long' => ['integer', 'long'],
+        'float' => ['number', 'float'],
+        'double' => ['number', 'double'],
+        'string' => 'string',
+        'date' => ['string', 'date'],
+        'datetime' => ['string', 'date-time'],
+        '\datetime' => ['string', 'date-time'],
+        'datetimeimmutable' => ['string', 'date-time'],
+        '\datetimeimmutable' => ['string', 'date-time'],
+        'datetimeinterface' => ['string', 'date-time'],
+        '\datetimeinterface' => ['string', 'date-time'],
+        'number' => 'number',
+        'object' => 'object',
+    ];
 
     public function __invoke(Analysis $analysis)
     {
         $refs = [];
-        if (!Generator::isDefault($analysis->openapi->components) && !Generator::isDefault($analysis->openapi->components->schemas)) {
-            foreach ($analysis->openapi->components->schemas as $schema) {
-                if (!Generator::isDefault($schema->schema)) {
-                    $refKey = $this->toRefKey($schema->_context, $schema->_context->class);
-                    $refs[$refKey] = OA\Components::ref($schema);
-                }
+        /** @var Definition $definition */
+        foreach ($analysis->swagger->definitions as $definition) {
+            if ($definition->definition) {
+                $refs[strtolower($definition->_context->fullyQualifiedName($definition->_context->class))] = '#/definitions/' . $definition->definition;
             }
         }
-
-        /** @var OA\Property[] $properties */
-        $properties = $analysis->getAnnotationsOfType(OA\Property::class);
-
-        foreach ($properties as $property) {
+        
+        $allProperties = $analysis->getAnnotationsOfType(Property::class);
+        /** @var Property $property */
+        foreach ($allProperties as $property) {
             $context = $property->_context;
-
-            if (Generator::isDefault($property->property)) {
+            // Use the property names for @SWG\Property()
+            if ($property->property === null) {
                 $property->property = $context->property;
             }
 
-            if (!Generator::isDefault($property->ref)) {
-                continue;
-            }
-
-            $typeAndDescription = $this->extractVarTypeAndDescription((string) $context->comment);
-
-            if (Generator::isDefault($property->type)) {
-                $this->augmentType($analysis, $property, $context, $refs, $typeAndDescription['type']);
-            } else {
-                $this->mapNativeType($property, $property->type);
-            }
-
-            if (Generator::isDefault($property->description) && $typeAndDescription['description']) {
-                $property->description = trim($typeAndDescription['description']);
-            }
-            if (Generator::isDefault($property->description) && $this->isRoot($property)) {
-                $property->description = $this->extractContent($context->comment);
-            }
-
-            if (Generator::isDefault($property->example) && ($example = $this->extractExampleDescription((string) $context->comment))) {
-                $property->example = $example;
-            }
-        }
-    }
-
-    protected function augmentType(Analysis $analysis, OA\Property $property, Context $context, array $refs, ?string $varType): void
-    {
-        // docblock typehints
-        if ($varType) {
-            $allTypes = strtolower(trim($varType));
-
-            if ($this->isNullable($allTypes) && Generator::isDefault($property->nullable)) {
-                $property->nullable = true;
-            }
-
-            $allTypes = $this->stripNull($allTypes);
-            preg_match('/^([^\[]+)(.*$)/', $allTypes, $typeMatches);
-            $type = $typeMatches[1];
-
-            // finalise property type/ref
-            if (!$this->mapNativeType($property, $type)) {
-                $refKey = $this->toRefKey($context, $type);
-                if (Generator::isDefault($property->ref) && array_key_exists($refKey, $refs)) {
-                    $property->ref = $refs[$refKey];
+            if (preg_match('/@var\s+(?<type>[^\s]+)([ \t])?(?<description>.+)?$/im', (string) $context->comment, $varMatches)) {
+                if ($property->description === null && isset($varMatches['description'])) {
+                    $property->description = trim($varMatches['description']);
                 }
-            }
+                if ($property->type === null) {
+                    preg_match('/^([^\[]+)(.*$)/', $this->stripNull(trim($varMatches['type'])), $typeMatches);
+                    $type = $typeMatches[1];
 
-            // ok, so we possibly have a type or ref
-            if (!Generator::isDefault($property->ref) && $typeMatches[2] === '' && !Generator::isDefault($property->nullable) && $property->nullable) {
-                $refKey = $this->toRefKey($context, $type);
-                $property->oneOf = [
-                    $schema = new OA\Schema([
-                        'ref' => $refs[$refKey],
-                        '_context' => new Context(['generated' => true], $property->_context),
-                    ]),
-                ];
-                $analysis->addAnnotation($schema, $schema->_context);
-                $property->nullable = true;
-            } elseif ($typeMatches[2] === '[]') {
-                if (Generator::isDefault($property->items)) {
-                    $property->items = $items = new OA\Items(
-                        [
-                            'type' => $property->type,
-                            '_context' => new Context(['generated' => true], $context),
-                        ]
-                    );
-                    $analysis->addAnnotation($items, $items->_context);
-                    if (!Generator::isDefault($property->ref)) {
-                        $property->items->ref = $property->ref;
-                        $property->ref = Generator::UNDEFINED;
-                    }
-                    $property->type = 'array';
-                }
-            }
-        }
-
-        // native typehints
-        if ($context->type && !Generator::isDefault($context->type)) {
-            if ($context->nullable === true) {
-                $property->nullable = true;
-            }
-            $type = strtolower($context->type);
-            if (!$this->mapNativeType($property, $type)) {
-                $refKey = $this->toRefKey($context, $type);
-                if (Generator::isDefault($property->ref) && array_key_exists($refKey, $refs)) {
-                    $this->applyRef($analysis, $property, $refs[$refKey]);
-                } else {
-                    if ($typeSchema = $analysis->getSchemaForSource($context->type)) {
-                        if (Generator::isDefault($property->format)) {
-                            $property->ref = OA\Components::ref($typeSchema);
-                            $property->type = Generator::UNDEFINED;
+                    if (array_key_exists(strtolower($type), static::$types)) {
+                        $type = static::$types[strtolower($type)];
+                        if (is_array($type)) {
+                            if ($property->format === null) {
+                                $property->format = $type[1];
+                            }
+                            $type = $type[0];
                         }
+                        $property->type = $type;
+                    } elseif ($property->ref === null && $typeMatches[2] === '') {
+                        $tmpKey = strtolower($context->fullyQualifiedName($type));
+                        $property->ref = array_key_exists($tmpKey, $refs) ? $refs[$tmpKey] : null;
+                    }
+                    if ($typeMatches[2] === '[]') {
+                        if ($property->items === null) {
+                            $property->items = new Items([
+                                'type' => $property->type,
+                                '_context' => new Context(['generated' => true], $context)
+                            ]);
+                            if ($property->items->type === null) {
+                                $tmpKey = strtolower($context->fullyQualifiedName($type));
+                                $property->items->ref = array_key_exists($tmpKey, $refs) ? $refs[$tmpKey] : null;
+                            }
+                        }
+                        $property->type = 'array';
                     }
                 }
             }
-        }
-
-        if (!Generator::isDefault($property->const) && Generator::isDefault($property->type)) {
-            if (!$this->mapNativeType($property, gettype($property->const))) {
-                $property->type = Generator::UNDEFINED;
+            if ($property->description === null) {
+                $property->description = $context->phpdocContent();
             }
         }
     }
 
-    protected function isNullable(string $typeDescription): bool
-    {
-        return in_array('null', explode('|', strtolower($typeDescription)));
-    }
-
-    protected function stripNull(string $typeDescription): string
+    /**
+     * @param string $typeDescription
+     *
+     * @return string
+     */
+    protected function stripNull($typeDescription)
     {
         if (strpos($typeDescription, '|') === false) {
             return $typeDescription;
@@ -163,22 +117,6 @@ class AugmentProperties implements ProcessorInterface
             }
             $types[] = $type;
         }
-
         return implode('|', $types);
-    }
-
-    protected function applyRef(Analysis $analysis, OA\Property $property, string $ref): void
-    {
-        if ($property->nullable === true) {
-            $property->oneOf = [
-                $schema = new OA\Schema([
-                    'ref' => $ref,
-                    '_context' => new Context(['generated' => true], $property->_context),
-                ]),
-            ];
-            $analysis->addAnnotation($schema, $schema->_context);
-        } else {
-            $property->ref = $ref;
-        }
     }
 }
