@@ -1,60 +1,52 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi\Processors;
+namespace Swagger\Processors;
 
-use OpenApi\Analysis;
-use OpenApi\Annotations as OA;
-use OpenApi\Context;
-use OpenApi\Generator;
+use Swagger\Annotations\Path;
+use Swagger\Logger;
+use Swagger\Context;
+use Swagger\Analysis;
 
 /**
- * Build the openapi->paths using the detected `@OA\PathItem` and `@OA\Operation` (`@OA\Get`, `@OA\Post`, etc).
+ * Build the swagger->paths using the detected @SWG\Path and @SWG\Operations (like @SWG\Get, @SWG\Post, etc)
  */
-class BuildPaths implements ProcessorInterface
+class BuildPaths
 {
     public function __invoke(Analysis $analysis)
     {
         $paths = [];
-        // Merge @OA\PathItems with the same path.
-        if (!Generator::isDefault($analysis->openapi->paths)) {
-            foreach ($analysis->openapi->paths as $annotation) {
-                if (empty($annotation->path)) {
-                    $annotation->_context->logger->warning($annotation->identity() . ' is missing required property "path" in ' . $annotation->_context);
-                } elseif (isset($paths[$annotation->path])) {
-                    $paths[$annotation->path]->mergeProperties($annotation);
-                    $analysis->annotations->detach($annotation);
-                } else {
-                    $paths[$annotation->path] = $annotation;
-                }
+        // Merge @SWG\Paths with the same path.
+        foreach ($analysis->swagger->paths as $annotation) {
+            if (empty($annotation->path)) {
+                Logger::notice($annotation->identity() . ' is missing required property "path" in ' . $annotation->_context);
+            } elseif (isset($paths[$annotation->path])) {
+                $paths[$annotation->path]->mergeProperties($annotation);
+                $analysis->annotations->detach($annotation);
+            } else {
+                $paths[$annotation->path] = $annotation;
             }
         }
 
-        /** @var OA\Operation[] $operations */
-        $operations = $analysis->unmerged()->getAnnotationsOfType(OA\Operation::class);
-
-        // Merge @OA\Operations into existing @OA\PathItems or create a new one.
+        // Merge @SWG\Operations into existing @SWG\Paths or create a new one.
+        $operations = $analysis->unmerged()->getAnnotationsOfType('\Swagger\Annotations\Operation');
         foreach ($operations as $operation) {
             if ($operation->path) {
                 if (empty($paths[$operation->path])) {
-                    $paths[$operation->path] = $pathItem = new OA\PathItem(
-                        [
-                            'path' => $operation->path,
-                            '_context' => new Context(['generated' => true], $operation->_context),
-                        ]
-                    );
-                    $analysis->addAnnotation($pathItem, $pathItem->_context);
+                    $paths[$operation->path] = new Path([
+                        'path' => $operation->path,
+                        '_context' => new Context(['generated' => true], $operation->_context)
+                    ]);
+                    $analysis->annotations->attach($paths[$operation->path]);
                 }
                 if ($paths[$operation->path]->merge([$operation])) {
-                    $operation->_context->logger->warning('Unable to merge ' . $operation->identity() . ' in ' . $operation->_context);
+                    Logger::notice('Unable to merge '.$operation->identity() .' in '.$operation->_context);
                 }
             }
         }
-        if ($paths) {
-            $analysis->openapi->paths = array_values($paths);
-        }
+        $analysis->swagger->paths = array_values($paths);
     }
 }
