@@ -1,163 +1,159 @@
-<?php declare(strict_types=1);
+<?php
+
+namespace Swagger;
 
 /**
- * @license Apache 2.0
- */
-
-namespace OpenApi;
-
-use OpenApi\Annotations as OA;
-use Symfony\Component\Yaml\Yaml;
-
-/**
- * Allows to serialize/de-serialize annotations from/to JSON.
+ * Class AnnotationDeserializer is used to deserialize a json string
+ * to a specific Swagger PHP Annotation class and vice versa.
  *
- * @see https://github.com/zircote/swagger-php
+ * @link https://github.com/zircote/swagger-php
  */
 class Serializer
 {
-    private static $VALID_ANNOTATIONS = [
-        OA\AdditionalProperties::class,
-        OA\Attachable::class,
-        OA\Components::class,
-        OA\Contact::class,
-        OA\Delete::class,
-        OA\Discriminator::class,
-        OA\Examples::class,
-        OA\ExternalDocumentation::class,
-        OA\Flow::class,
-        OA\Get::class,
-        OA\Head::class,
-        OA\Header::class,
-        OA\Info::class,
-        OA\Items::class,
-        OA\JsonContent::class,
-        OA\License::class,
-        OA\Link::class,
-        OA\MediaType::class,
-        OA\OpenApi::class,
-        OA\Operation::class,
-        OA\Options::class,
-        OA\Parameter::class,
-        OA\PathParameter::class,
-        OA\QueryParameter::class,
-        OA\CookieParameter::class,
-        OA\HeaderParameter::class,
-        OA\Patch::class,
-        OA\PathItem::class,
-        OA\Post::class,
-        OA\Property::class,
-        OA\Put::class,
-        OA\RequestBody::class,
-        OA\Response::class,
-        OA\Schema::class,
-        OA\SecurityScheme::class,
-        OA\Server::class,
-        OA\ServerVariable::class,
-        OA\Tag::class,
-        OA\Trace::class,
-        OA\Xml::class,
-        OA\XmlContent::class,
-    ];
+    const CONTACT = 'Swagger\Annotations\Contact';
+    const DEFINITION = 'Swagger\Annotations\Definition';
+    const DELETE = 'Swagger\Annotations\Delete';
+    const EXTERNALDOCUMENTATION = 'Swagger\Annotations\ExternalDocumentation';
+    const GET = 'Swagger\Annotations\Get';
+    const HEAD = 'Swagger\Annotations\Head';
+    const HEADER = 'Swagger\Annotations\Header';
+    const INFO = 'Swagger\Annotations\Info';
+    const ITEMS = 'Swagger\Annotations\Items';
+    const LICENSE = 'Swagger\Annotations\License';
+    const OPERATION = 'Swagger\Annotations\Operation';
+    const OPTIONS = 'Swagger\Annotations\Options';
+    const PARAMETER = 'Swagger\Annotations\Parameter';
+    const PATCH = 'Swagger\Annotations\Patch';
+    const PATH = 'Swagger\Annotations\Path';
+    const POST = 'Swagger\Annotations\Post';
+    const PROPERTY = 'Swagger\Annotations\Property';
+    const PUT = 'Swagger\Annotations\Put';
+    const RESPONSE = 'Swagger\Annotations\Response';
+    const SCHEMA = 'Swagger\Annotations\Schema';
+    const SECURITYSCHEME = 'Swagger\Annotations\SecurityScheme';
+    const SWAGGER = 'Swagger\Annotations\Swagger';
+    const TAG = 'Swagger\Annotations\Tag';
+    const XML = 'Swagger\Annotations\Xml';
 
-    protected static function isValidAnnotationClass(string $className): bool
+    private static $cachedNames;
+
+    private static function getDefinedNames()
     {
-        return in_array($className, self::$VALID_ANNOTATIONS);
+        if (static::$cachedNames === null) {
+            static::$cachedNames = [];
+            $reflection = new \ReflectionClass(__CLASS__);
+            static::$cachedNames = $reflection->getConstants();
+        }
+        return static::$cachedNames;
+    }
+
+    public static function isValidClassName($className)
+    {
+        return in_array($className, static::getDefinedNames());
     }
 
     /**
-     * Deserialize a string.
+     * Serialize.
+     *
+     * @param Annotations\AbstractAnnotation $annotation
+     * @return string
      */
-    public function deserialize(string $jsonString, string $className): OA\AbstractAnnotation
+    public function serialize(Annotations\AbstractAnnotation $annotation)
     {
-        if (!$this->isValidAnnotationClass($className)) {
-            throw new \Exception($className . ' is not defined in OpenApi PHP Annotations');
-        }
-
-        return $this->doDeserialize(json_decode($jsonString), $className, new Context(['generated' => true]));
+        return json_encode($annotation);
     }
 
     /**
-     * Deserialize a file.
+     * Deserialize a string
+     *
+     * @param $jsonString
+     * @param $className
+     *
+     * @return Annotations\AbstractAnnotation
+     *
+     * @throws \Exception
      */
-    public function deserializeFile(string $filename, string $format = 'json', string $className = OA\OpenApi::class): OA\AbstractAnnotation
+    public function deserialize($jsonString, $className)
     {
-        if (!$this->isValidAnnotationClass($className)) {
-            throw new \Exception($className . ' is not a valid OpenApi PHP Annotations');
+        if (!$this->isValidClassName($className)) {
+            throw new \Exception($className.' is not defined in Swagger PHP Annotations');
         }
+        return $this->doDeserialize(json_decode($jsonString), $className);
+    }
 
-        $contents = file_get_contents($filename);
-
-        $ext = pathinfo($filename, PATHINFO_EXTENSION);
-        if ('yaml' == $format || in_array($ext, ['yml', 'yaml'])) {
-            $contents = json_encode(Yaml::parse($contents));
+    /**
+     * Deserialize a file
+     *
+     * @param $filename
+     * @param $className
+     *
+     * @return Annotations\AbstractAnnotation
+     *
+     * @throws \Exception
+     */
+    public function deserializeFile($filename, $className = 'Swagger\Annotations\Swagger')
+    {
+        if (!$this->isValidClassName($className)) {
+            throw new \Exception($className.' is not defined in Swagger PHP Annotations');
         }
-
-        return $this->doDeserialize(json_decode($contents), $className, new Context(['generated' => true]));
+        $jsonString = file_get_contents($filename);
+        return $this->doDeserialize(json_decode($jsonString), $className);
     }
 
     /**
      * Do deserialization.
+     *
+     * @param \stdClass $c
+     * @param string    $class The class name of annotation.
+     *
+     * @return Annotations\AbstractAnnotation
      */
-    protected function doDeserialize(\stdClass $c, string $class, Context $context): OA\AbstractAnnotation
+    private function doDeserialize(\stdClass $c, $class)
     {
-        $annotation = new $class(['_context' => $context]);
-        foreach ((array) $c as $property => $value) {
+        $annotation = new $class([]);
+        foreach ($c as $property => $value) {
             if ($property === '$ref') {
                 $property = 'ref';
             }
 
             if (substr($property, 0, 2) === 'x-') {
-                if (Generator::isDefault($annotation->x)) {
-                    $annotation->x = [];
-                }
                 $custom = substr($property, 2);
                 $annotation->x[$custom] = $value;
             } else {
-                $annotation->{$property} = $this->doDeserializeProperty($annotation, $property, $value, $context);
+                $annotation->$property = $this->doDeserializeProperty($annotation, $property, $value);
             }
         }
-
-        if ($annotation instanceof OA\OpenApi) {
-            $context->root()->version = $annotation->openapi;
-        }
-
         return $annotation;
     }
 
     /**
      * Deserialize the annotation's property.
      *
-     * @param mixed $value
+     * @param Annotations\AbstractAnnotation $annotation
+     * @param string                         $property
+     * @param mixed                          $value
      *
      * @return mixed
      */
-    protected function doDeserializeProperty(OA\AbstractAnnotation $annotation, string $property, $value, Context $context)
+    private function doDeserializeProperty(Annotations\AbstractAnnotation $annotation, $property, $value)
     {
         // property is primitive type
         if (array_key_exists($property, $annotation::$_types)) {
-            return $this->doDeserializeBaseProperty($annotation::$_types[$property], $value, $context);
+            return $value;
         }
-
         // property is embedded annotation
-        // note: this does not support custom nested annotation classes
-        foreach ($annotation::$_nested as $nestedClass => $declaration) {
+        foreach ($annotation::$_nested as $class => $declaration) {
             // property is an annotation
             if (is_string($declaration) && $declaration === $property) {
-                if (is_object($value)) {
-                    return $this->doDeserialize($value, $nestedClass, $context);
-                } else {
-                    return $value;
-                }
+                return $this->doDeserialize($value, $class);
             }
 
             // property is an annotation array
             if (is_array($declaration) && count($declaration) === 1 && $declaration[0] === $property) {
                 $annotationArr = [];
                 foreach ($value as $v) {
-                    $annotationArr[] = $this->doDeserialize($v, $nestedClass, $context);
+                    $annotationArr[] = $this->doDeserialize($v, $class);
                 }
-
                 return $annotationArr;
             }
 
@@ -166,47 +162,13 @@ class Serializer
                 $key = $declaration[1];
                 $annotationHash = [];
                 foreach ($value as $k => $v) {
-                    $annotation = $this->doDeserialize($v, $nestedClass, $context);
-                    $annotation->{$key} = $k;
+                    $annotation = $this->doDeserialize($v, $class);
+                    $annotation->$key = $k;
                     $annotationHash[$k] = $annotation;
                 }
-
                 return $annotationHash;
             }
         }
-
-        return $value;
-    }
-
-    /**
-     * Deserialize base annotation property.
-     *
-     * @param array|string $type  The property type
-     * @param mixed        $value The value to deserialization
-     *
-     * @return array|OA\AbstractAnnotation
-     */
-    protected function doDeserializeBaseProperty($type, $value, Context $context)
-    {
-        $isAnnotationClass = is_string($type) && is_subclass_of(trim($type, '[]'), OA\AbstractAnnotation::class);
-
-        if ($isAnnotationClass) {
-            $isArray = strpos($type, '[') === 0 && substr($type, -1) === ']';
-
-            if ($isArray) {
-                $annotationArr = [];
-                $class = trim($type, '[]');
-
-                foreach ($value as $v) {
-                    $annotationArr[] = $this->doDeserialize($v, $class, $context);
-                }
-
-                return $annotationArr;
-            }
-
-            return $this->doDeserialize($value, $type, $context);
-        }
-
         return $value;
     }
 }
