@@ -1,15 +1,16 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * @license Apache 2.0
  */
 
-namespace OpenApi;
+namespace Swagger;
 
+use InvalidArgumentException;
 use Symfony\Component\Finder\Finder;
 
 /**
- * Convenient utility functions that don't neatly fit anywhere else.
+ * Convenient utility functions that don't neatly fit anywhere else
  */
 class Util
 {
@@ -24,9 +25,11 @@ class Util
      * and conform specifically to what is expected by functions like `exclude()` and `notPath()`.
      * In particular, leading and trailing slashes are removed.
      *
-     * @param array|string $basePaths
+     * @param string $fullPath
+     * @param string|array $basePaths
+     * @return string
      */
-    public static function getRelativePath(string $fullPath, $basePaths): string
+    public static function getRelativePath($fullPath, $basePaths)
     {
         $relativePath = null;
         if (is_string($basePaths)) { // just a single path, not an array of possible paths
@@ -39,45 +42,40 @@ class Util
                 }
             }
         }
-
         return !empty($relativePath) ? trim($relativePath, '/') : $fullPath;
     }
 
     /**
      * Removes a prefix from the start of a string if it exists, or null otherwise.
+     *
+     * @param string $str
+     * @param string $prefix
+     * @return null|string
      */
-    private static function removePrefix(string $str, string $prefix): ?string
+    private static function removePrefix($str, $prefix)
     {
         if (substr($str, 0, strlen($prefix)) == $prefix) {
             return substr($str, strlen($prefix));
         }
-
         return null;
     }
 
     /**
      * Build a Symfony Finder object that scans the given $directory.
      *
-     * @param array|Finder|string $directory The directory(s) or filename(s)
-     * @param null|array|string   $exclude   The directory(s) or filename(s) to exclude (as absolute or relative paths)
-     * @param null|string         $pattern   The pattern of the files to scan
-     *
-     * @throws \InvalidArgumentException
+     * @param string|array|Finder $directory The directory(s) or filename(s)
+     * @param null|string|array $exclude The directory(s) or filename(s) to exclude (as absolute or relative paths)
+     * @throws InvalidArgumentException
      */
-    public static function finder($directory, $exclude = null, $pattern = null): Finder
+    public static function finder($directory, $exclude = null)
     {
         if ($directory instanceof Finder) {
-            // Make sure that the provided Finder only finds files and follows symbolic links.
-            return $directory->files()->followLinks();
+            return $directory;
         } else {
             $finder = new Finder();
             $finder->sortByName();
         }
-        if ($pattern === null) {
-            $pattern = '*.php';
-        }
-
-        $finder->files()->followLinks()->name($pattern);
+        $finder->files()->followLinks()->name('*.php');
         if (is_string($directory)) {
             if (is_file($directory)) { // Scan a single file?
                 $finder->append([$directory]);
@@ -93,7 +91,7 @@ class Util
                 }
             }
         } else {
-            throw new \InvalidArgumentException('Unexpected $directory value:' . gettype($directory));
+            throw new InvalidArgumentException('Unexpected $directory value:' . gettype($directory));
         }
         if ($exclude !== null) {
             if (is_string($exclude)) {
@@ -103,52 +101,9 @@ class Util
                     $finder->notPath(Util::getRelativePath($path, $directory));
                 }
             } else {
-                throw new \InvalidArgumentException('Unexpected $exclude value:' . gettype($exclude));
+                throw new InvalidArgumentException('Unexpected $exclude value:' . gettype($exclude));
             }
         }
-
         return $finder;
-    }
-
-    /**
-     * Escapes the special characters "/" and "~".
-     *
-     * https://swagger.io/docs/specification/using-ref/
-     * https://tools.ietf.org/html/rfc6901#page-3
-     */
-    public static function refEncode(string $raw): string
-    {
-        return str_replace('/', '~1', str_replace('~', '~0', $raw));
-    }
-
-    /**
-     * Converted the escaped characters "~1" and "~" back to "/" and "~".
-     *
-     * https://swagger.io/docs/specification/using-ref/
-     * https://tools.ietf.org/html/rfc6901#page-3
-     */
-    public static function refDecode(string $encoded): string
-    {
-        return str_replace('~1', '/', str_replace('~0', '~', $encoded));
-    }
-
-    /**
-     * Shorten class name(s).
-     *
-     * @param array|object|string $classes Class(es) to shorten
-     *
-     * @return string|string[] One or more shortened class names
-     */
-    public static function shorten($classes)
-    {
-        $short = [];
-        foreach ((array) $classes as $class) {
-            $short[] = '@' . str_replace([
-                'OpenApi\\Annotations\\',
-                'OpenApi\\Attributes\\',
-                ], 'OA\\', $class);
-        }
-
-        return is_array($classes) ? $short : array_pop($short);
     }
 }
