@@ -1,8 +1,4 @@
-<?php declare(strict_types=1);
-
-/**
- * @license Apache 2.0
- */
+<?php
 
 namespace AnotherNamespace\Annotations;
 
