@@ -1,24 +1,16 @@
-<?php declare(strict_types=1);
-
-/**
- * @license Apache 2.0
- */
+<?php
 
 namespace AnotherNamespace;
 
-use OpenApi\Tests\Fixtures\ExpandClasses\Ancestor;
-
-use OpenApi\Annotations as OA;
-
 /**
- * @OA\Schema
+ * @SWG\Definition()
  */
-class Child extends Ancestor
+class Child extends \SwaggerFixtures\Ancestor
 {
+
     /**
      * @var bool
-     *
-     * @OA\Property
+     * @SWG\Property()
      */
     public $isBaby;
 }
