@@ -1,34 +1,27 @@
-<?php declare(strict_types=1);
+<?php
+namespace SwaggerFixtures;
 
 /**
- * @license Apache 2.0
- */
-
-namespace OpenApi\Tests\Fixtures;
-
-use OpenApi\Annotations as OA;
-
-/**
- * @OA\Info(title="Fixture for AugmentOperationTest", version="test")
+ * @SWG\Info(title="Fixture for AugmentOperationTest", version="test")
  */
 class UsingPhpDoc
 {
     /**
-     * Example summary.
+     * Example summary
      *
      * Example description...
      * More description...
      *
-     * @OA\Get(path="api/test1", @OA\Response(response="200", description="a response"))
+     * @SWG\Get(path="api/test1", @SWG\Response(response="200", description="a response"))
      */
     public function methodWithDescription()
     {
     }
 
     /**
-     * Example summary.
+     * Example summary
      *
-     * @OA\Get(path="api/test2", @OA\Response(response="200", description="a response"))
+     * @SWG\Get(path="api/test2", @SWG\Response(response="200", description="a response"))
      */
     public function methodWithSummary()
     {
